HLS: Implement XStreamInf tag
This CL implements the `XStreamInf` tag, which is used in multivariant
playlists to describe a variant. Currently only a subset of the
attributes of this tag are supported: those that are commonly used and
relevant to selecting an initial variant; other attributes are currently
ignored (which is in accordance with the HLS spec). Other attributes
will be added in future CLs once they are needed by the demuxer.

This CL adds relevant unit tests as well.

Bug: 1266991
Change-Id: I6c6dcbc31a1c7f6f677dfc8477c8acfc643f80a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3564113
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#989548}
NOKEYCHECK=True
GitOrigin-RevId: 58c7fc63d72a86d54a92323e262c25159942acc2