HLS: Refactor tag representation
This CL refactors a few design decisions I made early on about how tags
should be represented and parsed. The first being that tags were
identified based on their name as a prefix, which runs the problem of
potentially under-identifing or over-identifying tags. For example, if a
playlist contains "#EXTINF\n", that wouldn't be identified as an
`EXTINF` tag since its missing the colon. Similarly, if a new tag
"#EXT-X-GAP-FOO\n" were introduced, that would be identified as an
`EXT-X-GAP` tag with content "-FOO" and trigger a parsing error, rather
than being ignored as an unknown tag. The new design splits the tag name
and tag content on the colon, which is consistent with how all current
(and likely future) tags work, and makes tag content optional when the
colon isn't present. This also allows distinguishing between tags with
empty content, and tags missing content entirely.

Secondly, `kUnknownTagName` has been removed in favor of making
`TagItem::GetName()` return `absl::optional<TagName>`. This simplifies
code that can't handle unknown tags, and makes the process of reporting
unknown tag names simpler since content is not included in the reported
name.

Bug: 1266991
Change-Id: I945260b6e95ed67be2b8423ca292cfe82705d4f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3572217
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#991659}
NOKEYCHECK=True
GitOrigin-RevId: ee56fa571da0387a3feea15ccff4fd05e12f5fe8