media: Report UMA on first MediaFoundationCdm init
Currently we are seeing failures of MediaFoundationCdm initialization.
It's unclear whether those failures are happening right on the first
initialization, or only when we create many CDM instances. Adding a UMA
to only report on the first MediaFoundationCdm initialization to try
to help answer this question.

Bug: 1309741
Change-Id: I93bf1f1f7522e7d1cff53c8b5231c855f1518ede
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3587561
Reviewed-by: Evan Liu <evliu@google.com>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992987}
NOKEYCHECK=True
GitOrigin-RevId: 82b082eff4bc0d5ed4847fdc104784165fa9b218