HLS: Implement the EXT-X-TARGETDURATION tag
This CL implements the EXT-X-TARGETDURATION tag, which is required in
media playlists across all versions. This tag specifies an upper bound
on the length of media segments in a playlist. This tag is not
particularly relevant for VOD-style playlists, but is more useful when
performing playlist reloading.

Bug: 1266991
Change-Id: Ib1c7e55dc73fae714368fe780841a72fe78910f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3595299
Commit-Queue: Will Cassella <cassew@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994932}
NOKEYCHECK=True
GitOrigin-RevId: 3ff524ef35677d5b037b995917c6800b0a06861b