Avoid creating DRM framebuffer for video decoding.
Before this CL, the PlatformVideoFramePool (used by ChromeOS to allocate buffers for video decoding) used a GpuMemoryBufferFactoryNativePixmap to delegate the buffer allocation to Ozone. This factory keeps an internal cache of allocated NativePixmaps which has the effect of not having to call CreateNativePixmapFromHandle() in Ozone at the moment of creating a SharedImage using SharedImageBackingFactoryGLImage.

On the other hand, SharedImageBackingFactoryOzone unconditionally calls CreateNativePixmapFromHandle() (for NATIVE_PIXMAP GpuMemoryBuffers). If the PlatformVideoFramePool were to use a GpuMemoryBufferFactoryNativePixmap in this case, there would be two DRM framebuffers for each GpuMemoryBuffer: one would be created when allocating the buffer using the GpuMemoryBufferFactoryNativePixmap, and the other would be created when calling CreateNativePixmapFromHandle() during SharedImage creation. This makes some tests fail when switching to SharedImageBackingFactoryOzone because we exceed a threshold of allocated DRM framebuffers.

This CL prevents PlatformVideoFramePool from using a GpuMemoryBufferFactoryNativePixmap. The effect is that no DRM framebuffer will be created when creating a buffer for video decoding. It will only be created when creating the SharedImage.

Test: Tested video.MemCheck.h264_hw_switch and video.PlayDRM.* tests on crosfleet using volteer board.
Bug: b/213244572
Change-Id: Iaf69faa5919b48936e8ed91d8a3f36e34ef2f1bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3597211
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#996279}
NOKEYCHECK=True
GitOrigin-RevId: 791e99328e526262d88e4db5ad435475b45030bc