Add Audio Bitstream Passthrough Support For Secured Playback (Android)
This change allows an encrypted DTS audio bitstream to be passed
through to a connected Audio Sink device (e.g. AVR) for decoding.
Currently, un-encrypted audio bitstream passthrough works on Chromium
running on Android. Encrypted bitstream passthrough is blocked by a
MediaCodecUtil::CanDecode() check in cdm_message_filter_android.cc.
CanDecode() checks the MediaCodec library for a decoder that can decode
the bitstream. If not found, bitstream passthrough will be stopped.
This means that only Android devices with a capable decoder onboard
will be able to passthrough the bit-stream to the connected audio
device. The suggested improvement offered by this CL allows the
encrypted audio bitstream to be passthrough even when there is no
decoder in the MediaCodec library that can handle the bitstream.

Bug: 1311942
Change-Id: I91299f6e05ba60e857acc2685b22d5121dadba3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3562638
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Cedric Tio <cedric.tio@xperi.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#998099}
NOKEYCHECK=True
GitOrigin-RevId: c1fabe1d2d3579c822d98fa38ef8a22a7f0f6889