media: Create CDM instance when checking support
There are cases where the CDM factory says it supports playback, but
when we try to create the CDM it fails immediately. This CL add a step
to create a dummy CDM instance as part of the IsTypeSupported() logic
to detect this case.

Bug: 1309741
Change-Id: Id0f483d3daf211faafbda7cffbab88bb186ba6af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3617349
Reviewed-by: John Rummell <jrummell@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#999756}
NOKEYCHECK=True
GitOrigin-RevId: 3b496a74cab8f9471619a3bc423d3973bbc817b6