Each GLSurfaceEGL should be passed a pointer to a GLDisplayEGL.
In the constructor of GLSurfaceEGL, the surface should be passed a
pointer to the GLDisplayEGL that was used to create it. This pointer
should be held and referenced within the surface class.
For now, the GLDisplayEGL passed in is just the global GLDisplayEGL that
lives in gl_surface_egl.cc. Future CLs will pass in the correct
GLDisplayEGL that comes from the GLDisplayManager.

Bug: 1251724
Change-Id: I28bf7ac0be9d9f2985efbac05af70aa7cd302347
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3621554
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1000093}
NOKEYCHECK=True
GitOrigin-RevId: 35aed6c17512eda047f96054c0eff28320c434d6