HDR/Windows: Do not specify SDR white level in gfx::ColorSpace
This removes the places where we embed an SDR white level in a
gfx::ColorSpace. From now on, the SDR white level is an independent
parameter that is used only during tone mapping and conversion to an
SkColorSpace inside SkiaRenderer.

This allows specifying a different interpretation for SDR white level
when doing implicit color conversion (e.g, for import to 2D canvas,
WebGL, or WebGPU, where it is always 203 nits), and for platforms that
do not know their true luminance to select a reasonable synthetic value.

Bug: 1286076, 1274220
Change-Id: Ie92606e477bcf468dc65b48e48d87338d6eae92c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3630988
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1001290}
NOKEYCHECK=True
GitOrigin-RevId: 8f103a9c0033a1ec5630ac20f5e30141c0dd4944