Fix disabled video stream dropping EOS request token
OS: Win10

When playing video content via Media Foundation Renderer and switching
to a different tab will disable the stream. An EOS may be received at
this time for this stream. The EOS buffer should not cause
pending_sample_request_tokens_ to be dropped when stream is disabled
as this will cause media foundation to believe there is an outstanding
request when Media source does not and no longer request further sample
in the future.

Bug: 1324373
Change-Id: I4444af506f4d4c94cb47f5c948d41de9f9a68bdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3642596
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Daoyuan Li <daoyuanli@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1002706}
NOKEYCHECK=True
GitOrigin-RevId: c958eeb4c865ad133bd2b5a560cee736e22290d9