media/gpu/vaapiVEA: Release input frame before encoding
Originally VaapiVEA releases an input frame instance after encoding
for the frame is complete. This CL changes the releasing time to after
the frame is copied/imported to VA surface. A VA-API driver blocks any
access to the video frame after vaEndPicture() with the input VA
surface is executed until the encoding is complete, which is
vaMapBuffer() for a coded buffer returns.

This earlier releasing notifies a client the frame can be reused
earlier. This enables a client to process something while VaapiVEA
completes encoding, for instance, prepare the next frame to be
encoded.

Bug: b:229358029
Test: video_encode_accelerator_tests

Change-Id: Ia93e0100fbeb237eaf878afe87cfbaf603b48234
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3590337
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003047}
NOKEYCHECK=True
GitOrigin-RevId: 5dafa9467e134194dbcad8fd5fba5499db186458