[media] Improve handling of fallback buffers decoding
On initial decoder failure, DecoderStream copies all of the
|pending_buffers_| into |fallback_buffers_| and then proceeds to
decode all of the |fallback_buffers_|.

However, there may already be a pending read in the
DemuxerStream. Since decoders can have different maximum decode
requests, it is possible for the new decoder to have a smaller amount
of maximum decode requests than the previous one. When this happens,
the DemuxerStream may complete the pending read while a buffer from
the |fallback_buffers_| is being decoded, resulting in a DCHECK being
triggered in DecoderStream::DecodeInternal().

To work around this issue, this CL introduces a new state to
DecoderStream to indicate how many fallback buffers are currently
being decoded. We also check if we are hitting the corner case of the
new decoder having a smaller amount of maximum decode requests in
DemuxerStream::Decode() before calling
DemuxerStream::DecodeInternal(). This will also prevent from asking
for any more buffer from the DemuxerStream for as long as there are
fallback buffers being decoded.

Bug: b/225245735, 1324726, 1324732
Change-Id: I844ddc5dadfacde0dd68c23dbce9636ac16bafe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3553900
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Fabrice de Gans <fdegans@chromium.org>
Reviewed-by: Will Cassella <cassew@chromium.org>
Commit-Queue: Fabrice de Gans <fdegans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1005509}
NOKEYCHECK=True
GitOrigin-RevId: 1dcc9ac7063e2996c8cd2c93c7b1726ad88a1a7f