HLS: Implement the EXT-X-DISCONTINUITY-SEQUENCE tag
The EXT-X-DISCONTINUITY-SEQUENCE tag specifies an offset for the
discontinuity sequence number of the first media segment. These are
similar to media sequence numbers, except that they can be used to
synchronize media timelines across variants and renditions of a
multivariant playlist.

This CL implements the tag, as well as the process for assigning
discontinuity sequence numbers to media segments in a playlist. It
includes tests to verify this behavior matches the spec, and that it
does not interact with media sequence numbers unexpectedly.

Bug: 1266991
Change-Id: Id52fb0d5700fbf99e5a05345111c2b247b431c95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3639925
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1005549}
NOKEYCHECK=True
GitOrigin-RevId: 606f981d21790c9834ed4e603390b6aef0908d35