Eliminate base::LowerCaseEqualsASCII in /media
This CL is a behavior-preserving refactor. It replaces calls to
base::LowerCaseEqualsASCII with base::EqualsCaseInsensitiveASCII.

We are deprecating base::LowerCaseEqualsASCII because its interface is a
mild footgun, in that it requires the second parameter to be lowercase.

See the design doc for more details.
https://docs.google.com/document/d/18zWV0zN4oAqMZpond1Ut0QviHUeHCSwCo08mqThwHkQ/edit?usp=sharing

This CL was uploaded by git cl split.

R=sandersd@chromium.org

Bug: 1326529
Change-Id: I5be86b9ba258cb9e40f2f1ffa47449efc71e84f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3657580
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Auto-Submit: Dan McArdle <dmcardle@chromium.org>
Commit-Queue: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1006041}
NOKEYCHECK=True
GitOrigin-RevId: 56aea5c6e5c999fe4d4a22ed1aa458ce05f742fd