media/video: Add encoding mode to SupportedProfile
This CL adds an encoding mode bitmask to the SupportedProfile struct for
the VideoEncodeAccelerator. This enables users of the struct to check
for which encodingmode is supported by the profile. Right now, only CBR
is supported; VBR support will be added in follow-up CLs in accordance
with the designs linked below.

Main design (go/vbr-vea):
https://docs.google.com/document/d/1kofC1bNUgfJQ3DAlYuTlEkXhe86my0UrKzAXGHvnvgw/edit?usp=sharing
VA-API design:
https://docs.google.com/document/d/16kp6nfgMOvQWZRhB_4rzzVTIl6YhpSn46CvJ-jf628k/edit?usp=sharing

Bug: b:181797390
Change-Id: Ie04b81b85d16b62eea5f67d4fca3d7c8c4994589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3616902
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Clarissa Garvey <clarissagarvey@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1007049}
NOKEYCHECK=True
GitOrigin-RevId: 9d93d1105cc23cb2a6d459a628778ebb0959c5d0