HLS: Implement the EXT-X-BYTERANGE tag
This CL uses the previously implemented ByteRange type to implement the
EXT-X-BYTERANGE tag. This tag allows specifying that a media segment is
only a subset of the resource specified by its URI. The byterange
consists of a length and and offset, and the offset may be omitted if
the previous media segment is also a byterange of the same resource.
This CL also adds tests to validate that this works as expected.

Bug: 1266991
Change-Id: I2856913791f2fe549de1290835195cefb52c693e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3657257
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1007122}
NOKEYCHECK=True
GitOrigin-RevId: 6fa0934c74c42cd9092feaa49971c10464792e4b