HLS: Implement the EXT-X-PART-INF tag
This tag is an attribute-list of information about partial segments in
the playlist, and currently the only defined attribute is "PART-TARGET"
which is the target duration for partial segments. This tag is a
prerequisite to implementing partial segments. This CL adds unit tests
as necessary.

Bug: 1266991
Change-Id: I555ad451a14e951d066de458636b2714d5f85cc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3668518
Commit-Queue: Will Cassella <cassew@chromium.org>
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1007946}
NOKEYCHECK=True
GitOrigin-RevId: 5c7327e329011321379faad39ba29d7cf0c04dc4