Eliminate base::LowerCaseEqualsASCII in /media/gpu/vaapi
This CL is a behavior-preserving refactor. It replaces calls to
base::LowerCaseEqualsASCII with base::EqualsCaseInsensitiveASCII.

We are deprecating base::LowerCaseEqualsASCII because its interface is a
mild footgun, in that it requires the second parameter to be lowercase.

See the design doc for more details.
https://docs.google.com/document/d/18zWV0zN4oAqMZpond1Ut0QviHUeHCSwCo08mqThwHkQ/edit?usp=sharing

This CL was uploaded by git cl split.

R=jkardatzke@google.com

Bug: 1326529
Change-Id: Ibc2ac78d435aebaa4bd343a651eaef3ff35286c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3656653
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Auto-Submit: Dan McArdle <dmcardle@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1008330}
NOKEYCHECK=True
GitOrigin-RevId: 34337a16591a765b15fff5863d7561081e8f4a44