HLS: Use MediaPlaylist::CtorArgs structure
This CL replaces the ever-growing list of arguments to the
`media::hls::MediaPlaylist` constructor with a privately-defined
`CtorArgs` structure. The advantage to using this is the ability to
emulate "argument labels" for the constructor, which vastly improves
readability and makes it difficult to accidentally swap the order of
identically typed arguments between the constructor's declaration,
definition, and invocation.

Bug: 1266991
Change-Id: I4e3fb52f6febaec57f33f161c1cd3ffed3252eca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3689189
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Commit-Queue: Will Cassella <cassew@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1011268}
NOKEYCHECK=True
GitOrigin-RevId: ba6c570e187126ec25756e3474425b90956881d5