media/gpu/chromeos: remove check for PixelLayoutCandidate in Lacros
This CL removes a CHECK inside VideoDecoderPipeline for Lacros that
makes sure there is always a viable PixelLayoutCandidate. This check was
being triggered when testing hw accelerated video decoding for ARM
devices in Lacros (for instance, in jacuzzi). This check worked well
for intel/AMD devices since it allowed us not to create an image
processor. However, for some MediaTek devices we always need an
image processor. Therefore we are removing the CHECK for Lacros.

BUG=b:171813538
TEST=manual test on jacuzzi

Change-Id: I7ede37f1094adb86c8b5991e8558e93b231aef14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3663159
Commit-Queue: Pilar Molina Lopez <pmolinalopez@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012201}
NOKEYCHECK=True
GitOrigin-RevId: ff3c171d367608b951b13c286e3a040054fd9a10