media/gpu/vaapi/H264vaapiVED: Enable temporal layer encoding on AMD
VAConfigAttribEncMaxRefFrames reports the maximum number of
actually usable reference frames. We check the value is not less
than the picture length required for temporal layer encoding. This
is a stronger check as the maximum picture length can be more than
the maximum number of actually usable reference frames.
This CL removes the check because with a AMD driver the maximum
number of frames is 1 but the maximum picture length is 16.

The AMD driver used on ChromeOS is upreved to 22.0.3, so it supports
temporal layer encoding now. This CL enables the feature on ChromeOS
AMD devices.

Bug: b:199487660
Test: video.EncodeAccel.h264* on zork
Change-Id: Ie30d4d174ac780c98c040a8b84719108315bcc40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3288254
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012309}
NOKEYCHECK=True
GitOrigin-RevId: 09afbc8e9edee9332ae30f9ebf1cb11c3f85a9b3