StableVideoDecoderService: Implement OnVideoFrameDecoded().
This CL implements the StableVideoDecoderService::OnVideoFrameDecoded()
method which receives a notification from the underlying
mojom::VideoDecoder when a VideoFrame has been decoded.
StableVideoDecoderService::OnVideoFrameDecoded() only needs to forward
this call to the stable::mojom::VideoDecoderClient.

The tricky thing in this CL is that when writing the unit tests, I
realized that the StableVideoDecoderService should not call Construct()
directly on the mojom::VideoDecoder implementation. Instead, it should
do it through a mojo::Remote. The reason is that
mojom::VideoDecoder::Construct() takes a
mojo::PendingAssociatedRemote<mojom::VideoDecoderClient>. This
PendingAssociatedRemote must be sent over an existing mojo pipe.

Bug: b:195769334
Test: ./media_unittests --gtest_filter="StableVideoDecoderServiceTest.*" in amd64-generic VM.
Change-Id: I1e1bad067c005d0c7368f9ed73a48593ca3af903
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3654835
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Pilar Molina Lopez <pmolinalopez@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012319}
NOKEYCHECK=True
GitOrigin-RevId: 97cfc76dd653051040f33733878539929fa14022