[VideoCapture] Don't double configure the capture device on macOS.
The existing code was starting the device and then changing its
configuration, this almost doubles the startup time. A comment in
the code says it was done to ensure the configuration took effect.

macOS documentation[1] says instead that we should set the config
before starting the session with a block of beginConfiguration and
commitConfiguration calls.

Notably this documentation was just updated after Monday's WWDC,
previously it just said to hold lockForConfiguration when calling
startRunning... which did not appear to work: b/234196290

With this change resolution changes for the built-in camera on a
M1 MacBook Pro 14" go from ~550ms to ~350ms using the test page
from the bug. I saw no change in configuration times when using
an external Logitech C910 on an Intel Mac Mini.

This functionality is controlled by a new base::Feature called
"ConfigureCaptureBeforeStart" which is on by default (since the
new code is following Apple's recommended approach).

[1]: https://developer.apple.com/documentation/avfoundation/avcapturedevice/1389221-activeformat#discussion

Bug: 1324640
Change-Id: Id2da3340a0d36bf19879186cba97834214b8b697
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696666
Commit-Queue: Markus Handell <handellm@google.com>
Reviewed-by: Markus Handell <handellm@google.com>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1012874}
NOKEYCHECK=True
GitOrigin-RevId: d88b1513bba58af77a8cd5416f360176ca2b1c30