Add support for Intel HWDRM on ChromeOS in ARCVM
This mainly consists of adding a Mojo IPC call to extract the wrapped
key we need to inject into the video decoder. We have a call we use with
Chrome playback for this, but the daemon needs to know which environment
to get the key from so we need another call for Android.

This also adds support for switching cipher modes midstream on Intel.

This won't be enabled yet, there are other dependencies before this is
ready. GTS will be run on this before we completely enable it in the
build settings on ChromeOS.

The corresponding ChromeOS change is: crrev.com/i/3755973

BUG=b:203821839
TEST=ExoPlayer and Prime work on volteer w/ modified build config

Change-Id: Ie0e4939864642a4537ce863fe2caa3aac5c10508
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3684957
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1013624}
NOKEYCHECK=True
GitOrigin-RevId: 9db4817b7f94a311900dfc7c53fc99814fa4311c