[Region Capture] fix blue border in iframes
This patch changes the FSVCI to no longer output a region_capture_rect
whenever the region under capture is embedded in a separate renderer.
This effectively resolves a bug where the blue border is drawn in the
wrong location in this case.

Bug: 1327560
Change-Id: I99a2154f9f455e86df2aaa7151f84178e5202c1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3699030
Reviewed-by: Elad Alon <eladalon@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014101}
NOKEYCHECK=True
GitOrigin-RevId: 684d14c900b56ed8f21e3c59dfce9fafcafcef2c