HiDPI: fix heuristic for small capture sizes and odd aspect ratios
Add the original source size to frame metadata as input for the
heuristic. The content_rect in OnFrameCaptured is scaled to fit
the capture size, so it's not useful for scaling purposes if
the capture size is smaller than the original content size.

Also adjust the heuristic to use the same letterboxing approach
as the encoder for fitting the content rectangle in the capture
size.

Bug: 1324757
Change-Id: I57012a789e59e823fad92c96a5aeeb09105e081d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3698728
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1014309}
NOKEYCHECK=True
GitOrigin-RevId: b3a3859e45fbc943eb148c38ea12d25a58051e27