Plumb color space to gl::GLImageNativePixmap.
This CL plumbs the gfx::ColorSpace from the SharedImage system into the
gl::GLImageNativePixmap. This is important because until now,
gl::GLImageNativePixmap has assumed that YUV NativePixmaps can be
sampled as RGB assuming an encoding of BT.601 with limited range. This
is not correct in general (and we have found a case in the wild for
this in ChromeOS).

Bug: b:220336463, b:233667677
Test: None
Change-Id: I3eaa42612946ac3ac1bc3aa75e6edc67cfda19dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3662321
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1015234}
NOKEYCHECK=True
GitOrigin-RevId: 3e5555588156d74a22c924f246f54b9b8bbe3d3a