media: Process H265 PPS if none has been set yet
This was causing failures in CTS tests because it expects us to request
the buffers after only seeing the SPS/SPS and doesn't send an IDR frame
as well. We can safely do this if we have not selected a PPS yet.
Originally this code did do that, but then there are streams that have
multiple PPSs and activate different ones so we changed it to activate
it based on the PPS ID in IDR frames instead.

BUG=b:236242073
TEST=CtsMediaV2TestCases android.mediav2.cts.CodecDecoderTest PASS

Change-Id: I36324221348b737399d10f5a1c34eab10b851a36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3709983
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Cr-Commit-Position: refs/heads/main@{#1015496}
NOKEYCHECK=True
GitOrigin-RevId: bbe6390fb509343c2b50eed9de536a6787e2cb03