Protect ScopedAVPacket from misuse
With the previous definition, the compiler allowed a ScopedAVPacket to
be initialized with `new AVPacket`. In this case, the call to
av_packet_free() at destruction crashed randomly.

Now the compiler makes sure a ScopedAVPacket is only ever initialized
with the result of av_packet_alloc().

Change-Id: I15c130aa6ad8067b800ccca39f85a4be76a5feec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3704512
Commit-Queue: Wojciech Dzierżanowski <wdzierzanowski@opera.com>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1015836}
NOKEYCHECK=True
GitOrigin-RevId: 1f058b5d038e2048e96a4844bdb5385f16fd0a59