StableVideoDecoderService: Implement AddLogRecord().
This CL implements the StableVideoDecoderService::AddLogRecord() method
which receives a log event from the underlying mojom::VideoDecoder and
just needs to forward it to the remote stable::mojom::MediaLog.

The unit tests are expanded to cover the new code. Note that I had to
define the == operator for MediaLog in order to write the unit tests.

Bug: b:195769334
Test: ./media_unittests --gtest_filter="StableVideoDecoderServiceTest.*" in amd64-generic VM.
Change-Id: Ie0d257dbe6c0ab1accead4de31d749470fc7e05c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3682397
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Pilar Molina Lopez <pmolinalopez@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020126}
NOKEYCHECK=True
GitOrigin-RevId: 74538c849d6787c3a089b3421849ebcbd8e6e5dd