media: Don't strip YUV color space information for all RGB resources.
This CL fixes the VideoResourceUpdater so that it doesn't strip YUV
color space information from VideoFrames unless those VideoFrames are of
RGB format to begin with.

This is a follow-up to https://codereview.chromium.org/2738713003. The
intention of that CL was to avoid propagating YUV color space
information (like the YUV range) for VideoFrames that had been converted
from YUV to RGB by the time we produced the resource for the compositing
pipeline.

The particular condition my CL is fixing seems insufficient for that
purpose: typically, for hardware decoded frames in ChromeOS, the
resource type ends up being RGB even though the VideoFrame format is
NV12. That's because the YUV-to-RGB conversion is done at the time of
sampling the NV12 texture (if compositing) or at the time of scanning
out the NV12 buffer (if doing overlays). Therefore, we cannot strip the
YUV color space information in that case since that information might be
needed later in the pipeline. Besides, doing so would cause us to reject
the VideoFrame for overlays purposes (since CL:3662252).

Note: we keep the original condition but AND it with `IsRGB()` -- the
reason is that there are cases (such as Android) in which the frame
format is expected to be RGB but the resource type is expected to be
STREAM_TEXTURE. In those cases, we still don't want to discard the YUV
color space information because we want to forward things like the range
to SurfaceFlinger.

Bug: b:233667677, b:220336463, b:227956160
Test: play a 720p H.264 video from https://crosvideo.appspot.com to make sure it is promoted to overlay
Test: play a full-range video to make sure it's still not promoted to overlay (on volteer)
Test: graphics_HwOverlays.SkiaRenderer.video on volteer
Change-Id: Ic72287666cd91ab3ab1ac64cb5c57bb60f2ffc9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3749351
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1021982}
NOKEYCHECK=True
GitOrigin-RevId: ab2d51bf812491091e9adabae56609bef8eb283d